ai-worker/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
pillow==10.1.0
easyocr==1.7.1
openai==1.3.0
numpy==1.24.4
torch==2.1.0
torchvision==0.16.0
requests==2.31.0
pymongo==4.6.0
psycopg2-binary==2.9.9
